Organize test files for pages in 'offense' section into 'offense' subdirectory
