+++ cypress/e2e/felony/class-b/no-pros/offense/pages/f-offense-classb-4-cont.cy.ts
@@ -1,4 +1,4 @@
-import testNotSureModal from "../../../../utils/testNotSureModal";
+import testNotSureModal from "../../../../../utils/testNotSureModal";
 
 describe("Felony Class B No-Pros Fourth Page", () => {
   beforeEach(() => {
