+++ cypress/e2e/felony/class-b/no-pros/offense/pages/f-offense-classb-8-ineligible.cy.ts
@@ -1,4 +1,4 @@
-import testIneligiblePage from "../../../../utils/testIneligiblePage";
+import testIneligiblePage from "../../../../../utils/testIneligiblePage";
 
 describe("Felony Class B Ineligible Page", () => {
   it('should display ineligibility content due to RCW 9.94A.640(2)(b)', () => {
